refactor uploadMedia function into shared location
