+++ api/src/paths/activity.ts
@@ -1,23 +1,22 @@
 'use strict';
 
-import { ManagedUpload } from 'aws-sdk/clients/s3';
 import { RequestHandler } from 'express';
 import { Operation } from 'express-openapi';
 import { QueryResult } from 'pg';
 import { SQLStatement } from 'sql-template-strings';
 import { WRITE_ROLES } from '../constants/misc';
 import { getDBConnection } from '../database/db';
-import { ActivityPostRequestBody, IMediaItem, MediaBase64 } from '../models/activity';
+import { ActivityPostRequestBody } from '../models/activity';
 import geoJSON_Feature_Schema from '../openapi/geojson-feature-doc.json';
 import { getActivitySQL, IPutActivitySQL, postActivitySQL, putActivitySQL } from '../queries/activity-queries';
-import { uploadFileToS3 } from '../utils/file-utils';
 import { getLogger } from '../utils/logger';
+import { uploadMedia } from './media';
 
 const defaultLog = getLogger('activity');
 
-export const POST: Operation = [uploadMedia(), createActivity()];
+export const POST: Operation = [uploadMedia('activity'), createActivity()];
 
-export const PUT: Operation = [uploadMedia(), updateActivity()];
+export const PUT: Operation = [uploadMedia('activity'), updateActivity()];
 
 // Api doc common to both the POST and PUT endpoints
 const post_put_apiDoc = {
@@ -141,61 +140,6 @@ PUT.apiDoc = {
   description: 'Update an existing activity.',
   ...post_put_apiDoc
 };
-
-/**
- * Uploads any media in the request to S3, adding their keys to the request, and calling next().
- *
- * Does nothing if no media is present in the request.
- *
- * @returns {RequestHandler}
- */
-function uploadMedia(): RequestHandler {
-  return async (req, res, next) => {
-    defaultLog.debug({ label: 'activity', message: 'uploadMedia', body: req.body });
-
-    if (!req.body.media || !req.body.media.length) {
-      // no media objects included, skipping media upload step
-      return next();
-    }
-
-    const rawMediaArray: IMediaItem[] = req.body.media;
-
-    const s3UploadPromises: Promise<ManagedUpload.SendData>[] = [];
-
-    rawMediaArray.forEach((rawMedia: IMediaItem) => {
-      if (!rawMedia) {
-        return;
-      }
-
-      let media: MediaBase64;
-      try {
-        media = new MediaBase64(rawMedia);
-      } catch (error) {
-        defaultLog.debug({ label: 'uploadMedia', message: 'error', error });
-        throw {
-          status: 400,
-          message: 'Included media was invalid/encoded incorrectly'
-        };
-      }
-
-      const metadata = {
-        filename: media.mediaName || '',
-        description: media.mediaDescription || '',
-        date: media.mediaDate || '',
-        username: (req['auth_payload'] && req['auth_payload'].preferred_username) || '',
-        email: (req['auth_payload'] && req['auth_payload'].email) || ''
-      };
-
-      s3UploadPromises.push(uploadFileToS3(media, metadata));
-    });
-
-    const results = await Promise.all(s3UploadPromises);
-
-    req['mediaKeys'] = results.map((result) => result.Key);
-
-    next();
-  };
-}
 
 /**
  * Creates a new activity record.
