+++ api/src/paths/point-of-interest.ts
@@ -1,20 +1,19 @@
 'use strict';
 
-import { ManagedUpload } from 'aws-sdk/clients/s3';
 import { RequestHandler } from 'express';
 import { Operation } from 'express-openapi';
 import { SQLStatement } from 'sql-template-strings';
 import { WRITE_ROLES } from '../constants/misc';
 import { getDBConnection } from '../database/db';
-import { IMediaItem, MediaBase64, PointOfInterestPostRequestBody } from '../models/point-of-interest';
+import { PointOfInterestPostRequestBody } from '../models/point-of-interest';
 import geoJSON_Feature_Schema from '../openapi/geojson-feature-doc.json';
 import { postPointOfInterestSQL } from '../queries/point-of-interest-queries';
-import { uploadFileToS3 } from '../utils/file-utils';
 import { getLogger } from '../utils/logger';
+import { uploadMedia } from './media';
 
 const defaultLog = getLogger('point-of-interest');
 
-export const POST: Operation = [uploadMedia(), createPointOfInterest()];
+export const POST: Operation = [uploadMedia('point-of-interest'), createPointOfInterest()];
 
 POST.apiDoc = {
   description: 'Create a new point of interest.',
@@ -88,63 +87,6 @@ POST.apiDoc = {
     }
   }
 };
-
-/**
- * Uploads any media in the request to S3, adding their keys to the request, and calling next().
- *
- * Does nothing if no media is present in the request.
- *
- * TODO: make media handling an extension that can be added to different endpoints/record types
- *
- * @returns {RequestHandler}
- */
-function uploadMedia(): RequestHandler {
-  return async (req, res, next) => {
-    defaultLog.debug({ label: 'point-of-interest', message: 'uploadMedia', body: req.body });
-
-    if (!req.body.media || !req.body.media.length) {
-      // no media objects included, skipping media upload step
-      return next();
-    }
-
-    const rawMediaArray: IMediaItem[] = req.body.media;
-
-    const s3UploadPromises: Promise<ManagedUpload.SendData>[] = [];
-
-    rawMediaArray.forEach((rawMedia: IMediaItem) => {
-      if (!rawMedia) {
-        return;
-      }
-
-      let media: MediaBase64;
-      try {
-        media = new MediaBase64(rawMedia);
-      } catch (error) {
-        defaultLog.debug({ label: 'uploadMedia', message: 'error', error });
-        throw {
-          status: 400,
-          message: 'Included media was invalid/encoded incorrectly'
-        };
-      }
-
-      const metadata = {
-        filename: media.mediaName || '',
-        description: media.mediaDescription || '',
-        date: media.mediaDate || '',
-        username: (req['auth_payload'] && req['auth_payload'].preferred_username) || '',
-        email: (req['auth_payload'] && req['auth_payload'].email) || ''
-      };
-
-      s3UploadPromises.push(uploadFileToS3(media, metadata));
-    });
-
-    const results = await Promise.all(s3UploadPromises);
-
-    req['mediaKeys'] = results.map(result => result.Key);
-
-    next();
-  };
-}
 
 /**
  * Creates a new point of interest record.
